HOTT-1414: Dry up entities that have goods nomenclature
