+++ app/models/order_number/definition.rb
@@ -4,6 +4,7 @@ require 'null_object'
 class OrderNumber
   class Definition
     include ApiEntity
+    include HasGoodsNomenclature
 
     collection_path '/quotas'
 
@@ -38,14 +39,6 @@ class OrderNumber
 
     def id
       @id ||= quota_definition_sid
-    end
-
-    def all_goods_nomenclatures
-      measures
-        .map(&:goods_nomenclature)
-        .compact
-        .uniq(&:goods_nomenclature_item_id)
-        .sort_by(&:goods_nomenclature_item_id)
     end
 
     def geographical_areas
